Refactor lag correction for AutoRegImplementation
